pyproject.toml
[tool.poetry]
name = "simulacrum"
version = "0.1.dev0"
description = "declarative framework for in silico trials"
authors = ["Tyler M Kontra <tyler@havercene.io>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.7"
sqlalchemy = "^1.3.12"
pandas = "^0.25.3"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
pytest = "^5.3.2"
black = "^19.10b0"
rope = "^0.14.0"
pylint = "^2.4.4"
mypy = "^0.761"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
